rename to light_push, move class to separate file
