+++ packages/sdk/src/protocols/light_push/index.ts
@@ -1,0 +1,1 @@
+export { wakuLightPush } from "./light_push.js";
